sql script

+++ src/main/TIW.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS items (
   name VARCHAR(100) NOT NULL,
   description VARCHAR(255),
   cover_image VARCHAR(255) NOT NULL,
-  price INT DEFAULT 0,
+  price DOUBLE DEFAULT 0,
   sold BOOLEAN DEFAULT FALSE,
   created_by INT NOT NULL,
   FOREIGN KEY (created_by) REFERENCES users(id)
@@ -30,7 +30,7 @@ CREATE TABLE IF NOT EXISTS items (
 CREATE TABLE IF NOT EXISTS auctions (
       id INT AUTO_INCREMENT PRIMARY KEY,
       creator_id INT NOT NULL,
-      starting_price INT NOT NULL,
+      starting_price DOUBLE NOT NULL,
       min_increment INT NOT NULL,
       expiration DATETIME NOT NULL,
       closed BOOLEAN DEFAULT FALSE,
@@ -52,7 +52,7 @@ CREATE TABLE offers (
                         id INT AUTO_INCREMENT PRIMARY KEY,
                         auction_id INT NOT NULL,
                         user_id INT NOT NULL,
-                        offered_price DECIMAL(10,2) NOT NULL,
+                        offered_price DOUBLE NOT NULL,
                         offer_time DATETIME DEFAULT CURRENT_TIMESTAMP,
                         FOREIGN KEY (auction_id) REFERENCES auctions(id) ON DELETE CASCADE,
                         FOREIGN KEY (user_id) REFERENCES users(id)
@@ -62,7 +62,7 @@ CREATE TABLE offers (
 CREATE TABLE results (
                          auction_id INT PRIMARY KEY,
                          winner_id INT NOT NULL,
-                         final_price INT NOT NULL,
+                         final_price DOUBLE NOT NULL,
                          shipping_address VARCHAR(100) NOT NULL,
                          FOREIGN KEY (auction_id) REFERENCES auctions(id),
                          FOREIGN KEY (winner_id) REFERENCES users(id)
